calcNumUpdatesPerGroup: This function was allowing 0 as a valid number of updates per group. Changed to round up so there will always be at least one update per group.
I discovered other "bound" variables that were declared as ints. Changed them to doubles.

+++ cpp/src/hamiltonians/Random.cpp
@@ -56,7 +56,7 @@ consts::BondActionType Random::applyUpdate(Configuration& configuration, const L
 
 
 void Random::handleInsert(Configuration& configuration, const Lattice& lattice,
-    int groupLowerBound, int groupUpperBound) const {
+    double groupLowerBound, double groupUpperBound) const {
       
   int bondSize = chooseWeightedRandInt(bondTypeProps);
   double tauToInsert = chooseUnifRandDoubWithBounds(
@@ -108,7 +108,7 @@ void Random::handleInsert(Configuration& configuration, const Lattice& lattice,
 
 
 void Random::handleRemoval(Configuration& configuration, 
-    int groupLowerBound, int groupUpperBound) const {
+    double groupLowerBound, double groupUpperBound) const {
   std::vector<double> deletableTaus;
   std::set<double> taus = configuration.getTaus();
